template/pyproject.toml
[tool.poetry]
name = "dpy-template"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
discord-py = "^2.3.2"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
